create helper method to reduce repeated code
