+++ src/libs/utils.js
@@ -36,6 +36,32 @@ export const findPropertyValue = (propName, researcher) => {
   return isNil(prop) ? "" : prop.propertyValue;
 };
 
+export const getPropertyValuesFromUser = (user) => {
+  let researcherProps = {
+    academicEmail: user.email,
+    nihUsername: findPropertyValue(UserProperties.NIH_USERNAME, user),
+    linkedIn: findPropertyValue(UserProperties.LINKEDIN, user),
+    orcid: findPropertyValue(UserProperties.ORCID, user),
+    researcherGate: findPropertyValue(UserProperties.RESEARCHER_GATE, user),
+    department: findPropertyValue(UserProperties.DEPARTMENT, user),
+    division: findPropertyValue(UserProperties.DIVISION, user),
+    address1: findPropertyValue(UserProperties.ADDRESS1, user),
+    address2: findPropertyValue(UserProperties.ADDRESS2, user),
+    zipcode: findPropertyValue(UserProperties.ZIPCODE, user),
+    city: findPropertyValue(UserProperties.CITY, user),
+    state: findPropertyValue(UserProperties.STATE, user),
+    country: findPropertyValue(UserProperties.COUNTRY, user),
+    isThePI: findPropertyValue(UserProperties.IS_THE_PI, user),
+    havePI: findPropertyValue(UserProperties.HAVE_PI, user),
+    piName: findPropertyValue(UserProperties.IS_THE_PI, user) === "true" ? user.displayName : findPropertyValue(UserProperties.PI_NAME, user),
+    piEmail: findPropertyValue(UserProperties.IS_THE_PI, user) === "true" ? user.email : findPropertyValue(UserProperties.PI_EMAIL, user),
+    pubmedID: findPropertyValue(UserProperties.PUBMED_ID, user),
+    scientificURL: findPropertyValue(UserProperties.SCIENTIFIC_URL, user)
+  };
+
+  return researcherProps;
+};
+
 export const applyHoverEffects = (e, style) => {
   forEach(style, (value, key) => {
     e.target.style[key] = value;
